fix(independence): restore assumptions accordion bar, drop sidebar layout
