+++ src/components/features/independence/index.ts
@@ -40,7 +40,7 @@ export { default as SustainableSpendingCard } from "./SustainableSpendingCard"
 export { default as SpendableAtIndependenceCard } from "./SpendableAtIndependenceCard"
 export { default as SetDateOfBirthNotice } from "./SetDateOfBirthNotice"
 export { default as PlanFindingsCard } from "./PlanFindingsCard"
-export { default as AssumptionsPanel } from "./AssumptionsPanel"
+export { default as ScenarioBar } from "./ScenarioBar"
 export { default as StrategyGaugesStrip } from "./StrategyGaugesStrip"
 export { default as PensionGauge } from "./PensionGauge"
 export { default as TimelineTabContent } from "./TimelineTabContent"
